# core/acowork-pm/schemas/pm-openapi-3.1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "acowork-pm REST API",
    "version": "0.1.0",
    "description": "项目 / 任务 / 附件管理服务 REST 契约。\n\n- **Base URL**：生产经 Gateway 反向代理 `{gw}/api/pm/...` → 剥 `/api/pm` 前缀转发到 pm 服务内部路径（不带 `/api` 前缀）；开发独立 serve 时路径不带 `/api` 前缀。\n- **鉴权**：复用 Gateway Desktop 会话鉴权（Bearer Token，见 `docs/protocols/zh/http.md` §1）。\n- **Actor 标识**：写操作与 claim/submit/review 依赖 HTTP header `X-Actor`（Gateway 注入当前用户/Agent ID）。\n- **错误格式**：统一 `{\"error\": {\"code\": \"...\", \"message\": \"...\"}}`，错误码表见 `schemas/README.md` §3 与 `src/error.rs`。\n- **契约权威**：本文件与 `src/api/routes.rs` 路由注册、`src/types.rs` 类型定义一一对应；设计与本契约冲突时以本契约为准（实现态）。"
  },
  "paths": {
    "/api/pm/projects": {
      "get": {
        "summary": "项目列表",
        "operationId": "listProjects",
        "responses": {
          "200": {
            "description": "项目列表",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Project" } } } }
          }
        }
      },
      "post": {
        "summary": "新建项目",
        "operationId": "createProject",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateProject" } } }
        },
        "responses": {
          "200": { "description": "创建成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Project" } } } }
        }
      }
    },
    "/api/pm/projects/{pid}": {
      "parameters": [{ "$ref": "#/components/parameters/Pid" }],
      "get": {
        "summary": "项目详情",
        "operationId": "getProject",
        "responses": {
          "200": { "description": "项目详情", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Project" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "summary": "更新项目",
        "operationId": "updateProject",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateProject" } } }
        },
        "responses": {
          "200": { "description": "更新成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Project" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "summary": "删除项目（级联任务，需二次确认）",
        "operationId": "deleteProject",
        "parameters": [
          { "name": "cascade", "in": "query", "required": false, "description": "强制级联删除所有任务（默认 false，仍有任务时返回 409）", "schema": { "type": "boolean", "default": false } }
        ],
        "responses": {
          "204": { "description": "删除成功" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "Pid": { "name": "pid", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/ProjectId" } },
...
    },
    "responses": {
...
      "NotFound": { "description": "资源不存在", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorBody" } } } },
      "Conflict": { "description": "结构冲突（环 / 依赖未满足 / 删除前仍有子项）", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorBody" } } } }
    },
    "schemas": {
      "ProjectId": { "type": "string", "pattern": "^p-[a-zA-Z0-9-]+$", "minLength": 3, "maxLength": 64, "example": "p-3f2a9c1e" },
...
      "DateTimeUtc": { "type": "string", "format": "date-time", "example": "2025-03-01T08:00:00Z" },
      "ProjectStatus": { "type": "string", "enum": ["active", "archived", "completed"] },
...
      "Project": {
        "type": "object",
        "required": ["id", "title", "status", "created_by", "created_at", "updated_at"],
        "properties": {
          "id": { "$ref": "#/components/schemas/ProjectId" },
          "title": { "type": "string" },
          "description": { "type": "string", "default": "" },
          "status": { "$ref": "#/components/schemas/ProjectStatus" },
          "created_by": { "type": "string" },
          "created_at": { "$ref": "#/components/schemas/DateTimeUtc" },
          "updated_at": { "$ref": "#/components/schemas/DateTimeUtc" },
          "metadata": { "type": "object", "additionalProperties": true }
        }
      },
      "CreateProject": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string", "default": "" },
          "metadata": { "type": "object", "additionalProperties": true }
        },
        "additionalProperties": false
      },
      "UpdateProject": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "status": { "$ref": "#/components/schemas/ProjectStatus" },
          "metadata": { "type": "object", "additionalProperties": true }
        },
        "additionalProperties": false
      },
...
      "ErrorBody": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string", "description": "机器可读错误码（snake_case），见 schemas/README.md §3" },
              "message": { "type": "string", "description": "人类可读错误信息" }
            }
          }
        }
      }
    }
  }
}